requirements.txt
python-telegram-bot==20.7
fastapi>=0.128.0
uvicorn>=0.40.0
gigachat==0.2.0b2
python-dotenv>=1.0.1
pydantic>=2.8.0